move user and auth type to dto type file
